refactor: sha256 pinning
