+++ src/codeLens/events/onUpdateDependencyClick.ts
@@ -22,11 +22,25 @@ export class OnUpdateDependencyClick {
 
     const pkg = codeLens.pkgResult.value
     const edit = new vscode.WorkspaceEdit()
-    edit.replace(
-      codeLens.documentUrl,
-      codeLens.replaceRange,
-      formatWithExistingLeading(codeLens.dependency.specifier, pkg.version),
-    )
+
+    if (pkg.format === 'github-actions-workflow' && pkg.alias) {
+      // GitHub Actions: replace specifier (and any existing comment) with SHA256 alias and tag as comment
+      const document = await vscode.workspace.openTextDocument(codeLens.documentUrl)
+      const lineText = document.lineAt(codeLens.replaceRange.start.line).text
+      const lineEnd = lineText.trimEnd().length
+      const extendedRange = new vscode.Range(
+        codeLens.replaceRange.start,
+        new vscode.Position(codeLens.replaceRange.start.line, lineEnd),
+      )
+      edit.replace(codeLens.documentUrl, extendedRange, `${pkg.alias} # ${pkg.version}`)
+    } else {
+      edit.replace(
+        codeLens.documentUrl,
+        codeLens.replaceRange,
+        formatWithExistingLeading(codeLens.dependency.specifier, pkg.version),
+      )
+    }
+
     await vscode.workspace.applyEdit(edit)
   }
