+++ src/package/gem.ts
@@ -1,9 +1,32 @@
 import camelcaseKeys from 'camelcase-keys'
 import urlJoin from 'url-join'
+import { z } from 'zod'
 
-import { GemSchema, GemVersionsSchema, GemVersionsType, PackageType } from '@/schemas'
+import { PackageType } from '@/schemas'
 
 import { AbstractPackageClient } from './abstractClient'
+
+export const GemVersionSchema = z.object({
+  number: z.string(),
+})
+
+export const GemVersionsSchema = z.array(GemVersionSchema)
+
+export type GemVersionsType = z.infer<typeof GemVersionsSchema>
+
+export const GemSchema = z.object({
+  version: z.string(),
+  info: z.string(),
+  homepageUri: z.string(),
+})
+
+export const GemPackageSchema = z.object({
+  name: z.string(),
+  requirements: z.string().nullish(),
+})
+
+export type GemType = z.infer<typeof GemSchema>
+export type GemPackageType = z.infer<typeof GemPackageSchema>
 
 export class GemClient extends AbstractPackageClient {
   constructor(privateSource?: string) {
